+++ Project/Renderer/DynamicDescriptorPool.h
@@ -19,8 +19,8 @@ public:
 private:
 	ID3D12Device* m_pDevice = nullptr;
 	ID3D12DescriptorHeap* m_pDescriptorHeap = nullptr;
-	UINT m_AllocatedDescriptorCount = 0;
-	UINT m_MaxDescriptorCount = 0;
+	UINT64 m_AllocatedDescriptorCount = 0;
+	UINT64 m_MaxDescriptorCount = 0;
 	UINT m_CBVSRVUAVDescriptorSize = 0;
 	D3D12_CPU_DESCRIPTOR_HANDLE m_CPUDescriptorHandle = { 0xffffffffffffffff, };
 	D3D12_GPU_DESCRIPTOR_HANDLE m_GPUDescriptorHandle = { 0xffffffffffffffff, };
